+++ app/utils/bubble-action.js
@@ -1,0 +1,5 @@
+export default function bubbleAction ( proxyTo ) {
+  return function () {
+    this.get(proxyTo).apply(this, arguments);
+  };
+}